[keyserver][lib] Simplify remaining thread responders following deprecation of old codeVersions
Summary:
Doing these together because they share the same types and validators.

This is the last of four diffs that simplify the code, types, and validators for endpoints mutating threads. These can be simplified following the deprecation of older `codeVersion`s in the first diff of this stack.

I finally update the code in `thread-permission-updaters.js` that I mentioned [here](https://linear.app/comm/issue/ENG-108#comment-8b4a79c8), which is sort of the whole point of this stack. I'm hoping this change will improve performance of most of these endpoints.

Depends on D8508

Test Plan: I tested this whole stack together by playing around with the web app and making sure thread mutation endpoints seemed to work correctly

Reviewers: michal, kamil

Subscribers: