[web] Inherit min-height property for media elements
Summary:
Fix for [ENG-4400](https://linear.app/comm/issue/ENG-4400/media-elements-lose-min-height-property).
This diff fixes the issue when the rendered height of the image (or video) is less than 50px. In such cases, the image doesn't cover the container completely,
resulting in the absence of borders.

Test Plan: Upload some photos and videos, resize browser window and check if image min-height is maintained.

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8518