[web] Fix closing MultimediaModal on ESC press bug
Summary:
Fix for [ENG-4327](https://linear.app/comm/issue/ENG-4327/allow-esc-to-close-imagemodal-on-web).

This diff solves the issue, where closing MultimediaModal with the ESC key was impossible.

Test Plan: In chat, click on the image and when MultimediaModal is shown, check if ESC key press closes it. Tested on unencrypted and encrypted images.

Reviewers: bartek, tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8498