[keyserver] Correctly set memberships last_message
Summary:
[ENG-4101](https://linear.app/comm/issue/ENG-4101)

Previously when creating a new thread the `memberships.last_message` was set to 1 (or 0 if it was supposed to be unread). We should set it to correct value, because after D8492 we depends on this value t always make sense. One problem is that if a user doesn't have a visiblity into a subthread, they don't see the subthread creation message so we need to handle this. Another problem is that we can't just query the db, because some of the membership rows are yet to be created, so we need to handle this case too.

Test Plan:
- Register a new user, check if their genesis last message is set to the last message in genesis
- Create a secret subthread, add a user, check that their last message is set to the message directly before
- Create an open subthread, add a user, check that their last message is set to the subthread-creation-message
(in the last two cases the last_message is then overriden by joined-chat message)

Reviewers: kamil, inka

Subscribers: