[lib, keyserver, native, web] Remove the cookie field from redux
Summary:
https://linear.app/comm/issue/ENG-4377/refactor-cookie-field-in-web-redux
Removing the cookie field from redux. Adding a migration. Undoing a hack in `cookieSelector`.

Test Plan:
ran `yarn flow-all`. Since cookie was not being used since the previous diff, this should be enough

EDIT:
tested that the cookie was migrated on native:
- checked out master, opened the app, logged the cookie
- closed the app
- checked out this diff, opened the app, logged the cookie

On master, rehyderated keys included `cookie`. On this diff, rehydrated keys don't include 'cookie', but include the `keyserverStore`. Checked that `cookie` that was logged on master is equal to the `keyserverStore.keyserverInfos['256'].cookie` on this diff

Reviewers: michal, kamil

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8536