[native] Display an action sheet with the 'edit role' option
Summary: The first option we display to the user is an "Edit role" item, where clicking it will take them to the role name / permissions screen. This diff just handles the action sheet, but the next one will
handle responding to pressing the 'Edit role' option. Down the line, I will also be adding the 'Delete' option to the action sheet.

[[ https://linear.app/comm/issue/ENG-4368/display-an-action-sheet-with-the-edit-role-option | ENG-4368 ]].

Depends on D8522

Test Plan: Verified that on iOS and Android, the action sheet looks as expected.

Reviewers: atul, ginsu

Subscribers: