[keyserver] Handle offensive words in invite links
Summary:
Check if an invite link contains abusive words and return an error if that's the case.

Depends on D8494

Test Plan:
Tried to create a link consisting of just an offensive word and an error was returned.
Tried to create a link with a string consisting of an offensive word with some prefix - a link was created correctly. This isn't ideal, but handing it correctly might be challenging: e.g. `invite/class` sounds like a proper link, but simply checking if it contains an offensive substring would forbid it.

Reviewers: kamil, inka, ashoat

Differential Revision: https://phab.comm.dev/D8527