[web] switch code from `sql.js` API to `.wasm` database module
Summary:
This is very important diff - it starts using `.wasm` API from `SQLiteQueryExecutor` instead of `sql.js`.

We do not have to migrate this content, we will simply take old content from indexedDB and put this to database module (schema is the same).

Depends on D8559

Test Plan: Test everything end-to-end.

Reviewers: michal, tomek

Reviewed By: michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8560