[desktop] Add context menu
Summary:
[ENG-4407](https://linear.app/comm/issue/ENG-4407/no-way-to-right-click-in-electron#comment-0e8041e9)

Adds a right click context menu to the electron app. Uses `electron-context-menu` for that, because it comes with some nice defaults and window handling.

The [github page readme](https://github.com/sindresorhus/electron-context-menu) contains a nice explanation of available options, so if anyone thinks we should change the list I'm open to that.

Adding ashoat as there's a new dependency

Test Plan:
Right click on text and images and check if context menu is displayed:

I checked if all visibile options work as expected.

On windows it should work the same way, it will just show an additional "Select All" option.

Reviewers: kamil, inka, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8573