[native] introduce functionality to send a reaction by pressing on the reaction pill.
Summary:
To improve the user experience of the `InlineEngagement` component we are updating the pressing behavior for our reaction pills. Before whenever a user would press on a reaction, the `MessageReactionsModal` would slide in, but now if a user presses a reaction they will react/unreact to that message depending on whether or not they have already reacted to that message already. To view the `MessageReactionsModal` the user now needs to long press on the reaction pill (a similar pattern with other chat apps such as discord and slack)

Depends on D8576

Test Plan:
Please watch the demo below

{F645787}

Reviewers: atul, kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8578