[services][terraform] Set up ECS cluster
Summary:
This diff introduces resources needed for ECS Cluster that is able to launch tasks on EC2 instances.

The cluster was created in AWS Console and imported to Terraform, along with default services namespace.

Additionally, the following resources were needed to run tasks on EC2:
- Launch template with some custom instance configuration
- Autoscaling group controlled by ECS capacity provider

Depends on D8582.

Test Plan: This configuration is already working in production for three services. Specific service configurations will be added in subsequent diffs.

Reviewers: jon, varun, michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8583