[services][terraform] Set up electron-update ECS service
Summary:
This diff introduces Terraform configuration that is needed to bootstrap the electron-update service.
It is deployed as a ECS Fargate service - this diff more or less the deployment configuration for services.

A few notes:
- The task definition, ecs-service and target-group were created using Terraform
- Existing load balancer was imported into Terraform state and the ECS target group was linked to it
- A test listener was added to be able to reach the ECS service directly. I'll delete it once we redirect 100% traffic to the ECS service.

Depends on D8583

Test Plan: The service is already deployed to production and handles most of real network traffic.

Reviewers: michal!, jon, varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8585