[web] Display a Roles option in the community menu
Summary:
If the user has `CHANGE_ROLE` permissions, we want to allow them to see the 'Roles' option in the community actions menu. Implementation will come in the following diffs, this just gets the icon /
option shown.

[[ https://linear.app/comm/issue/ENG-4425/display-a-roles-option-in-the-community-menu | ENG-4425 ]]

Depends on D8591

Test Plan:
Checked that the option shows for my local admin user only

{F648295}

Reviewers: atul, ginsu

Reviewed By: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8592