[keyserver/lib] Enable the option to delete custom roles
Summary:
This diff handles the flow for deleting custom roles. There are two cases to consider here when a role is deleted: (1) if there are no members assigned to the role, just delete it. (2) If there are members assigned to the role, then we need to display in the Alert message that they will be assigned to the default role (I did not hardcode the role name 'Members' since we're allowing it to be renamed, so instead I pull the default role name from the threadInfo).

A lot of the code in this diff is the scaffolding for the deletion API, so I decided that covering it all in one diff would be more logical and easier to review.

[[ https://linear.app/comm/issue/ENG-4389/enable-the-option-to-delete-custom-roles | ENG-4389 ]]

Depends on D8564

Test Plan:
Confirmed that role deletion works successfully for both cases (1) and (2).

{F645347}

Reviewers: atul, ginsu, ashoat!

Reviewed By: ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8574