[services][terraform] Set up blob ECS service
Summary:
This diff demonstrates a ECS service that is hosted on a EC2 instance.
Blob Service exposes both HTTP and gRPC endpoints so it was a good candidate for this demo.

Depends on D8583

Test Plan: ECS-based blob service is up and running in production. Tested HTTPS endpoints via Postman and gRPC (SSL) endpoints with BloomRPC.

Reviewers: varun, jon, michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8598