[CommCoreModule] implement C++ DraftStore
Summary: Grouping logic in one class to make it cleaner and reuse some code.

NOTE: I will wait with creating diffs for the rest of the stores until this is accepted because rest will be very similar to this one.

Depends on: D8603

Test Plan: Build iOS and Android and play with drafts to make sure functionality still works.

Reviewers: michal, jon, atul, tomek