[keyserver][lib] Block users from creating reserved invite links
Summary: A user should not be able to create a link which is included in a list of reserved usernames. The copy is similar to a message that is shown when a user tries to create an account with reserved username: `/native/account/registration/registration-server-call.js`, but I'm not sure if we should mention support email (the user is already registered and can contact us on Comm).

Depends on D8527

Test Plan: Try to create `/comm` invite link - a message should be displayed.

Reviewers: kamil, inka, ashoat