[native] Navigate to the role creation screen with pre-populated fields for editing roles
Summary:
This diff handles the edit roles action on the client-side. Keyserver support for editing roles will be a minimal change and will come in the following diff. I was able to re-use the create-roles
screens/flow here, and passed in the existing permissions and role name to pre populate the fields.

Test Plan: Confirmed that the navigation works as expected and the correct permissions get selected on navigation.

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8524