[native] introduce DummyInlineEngagementNode
Summary: This next part of the stack will introduce the logic for updating the chat item height measurer to include the inline engagement in the message measurement. The first diff introduces the dummy inline

Test Plan: Confirmed that the incorrect height measuremnt log was no longer in the console. Also double checked in the inspector that I was getting the correct height measurments for both dummy nodes

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8610