[native] introduce DummyInlineEngagementNode
Summary:
This next part of the stack will introduce the logic for updating the chat item height measurer to include the inline engagement in the message measurement. The order of operations for this next part of the stack will be correctly measuring the robotext message, then a text message, then finally a multimedia message.

The first diff introduces the dummy inline engagement component that will be a part of the `heightMeasurerDummy` which will get passed to the `NodeHeightMeasurer`. Subsequent diffs will handle updating the `heightMeasurerKey` function.

Depends on D8609

Test Plan:
Confirmed that the incorrect height measurement log was no longer in the console. Also double checked in the inspector that I was getting the correct height measurements for both dummy nodes. I will also be continually testing this component in subsequent diffs

Before:
{F652127}

After:
{F652128}

Reviewers: atul, kamil, tomek, ashoat

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8610