[lib/native/web] update getMessageLabel parameters
Summary:
Small refactor diff so that we can use `getMessageLabel` in a subsequent diff. In this subsequent diff we will use this function, but we will only have access to the `threadID` from the `messageInfo`. Since we are only using the `threadID` in this function, it seems better to only pass the `threadID` instead of the whole `threadInfo`

Depends on D8628

Test Plan: `flow` and everything still works as expected

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8629