[native] update dummyNodeForTextMessageHeightMeasurement to use dummy inline engagement component
Summary:
This diff adds the dummy inline engagement component to the text message dummy node, so that we can correctly measure text messages with an inline engagement. A subsequent diff will handle updating the `heightMeasurerKey` for text messages

Depends on D8629

Test Plan:
Confirmed that the incorrect height measurement log was no longer in the console. Also double checked in the inspector that I was getting the correct height measurements for both dummy nodes

Before:
{F653906}

After:
{F653907}

Reviewers: atul, kamil, ashoat, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8630