[native] update heightMeasurerKey function to consider inline engagement for text messages
Summary:
This next diff updates the heightMeasurerKey function so that we remeasure a text message whenever there is a change with the inline engagement for that text message. In this diff we want to also consider changes to sidebar text and the reactions, and the edited label since these can all potentially affect the height of the inline engagement node

Depends on D8630

Test Plan:
I tested modifying the height of the inline engagement for text messages, and when I did this I no longer got incorrect height measurement logs from the console. Also confirmed that there were no incorrect height measurements for android

When modifying the height I tried various things:
- introducing/removing reactions (demo below)
- introducing a sidebar to the inline engagement (pushing some of the reactions to a new row)
- introducing the edited label to the inline engagement  (pushing some reactions to a new row)

With all of these tests, I did not get any logs about incorrect height measurements.

Before:
{F654088}

After:
{F654094}

Reviewers: atul, kamil, ashoat, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8632