[keyserver] Use custom url on intermediate invite link page
Summary:
Clicking universal links (https) in other browsers than Safari doesn't open the app. But using `comm://` scheme should work correctly.

https://linear.app/comm/issue/ENG-4089/invite-links-dont-work-on-prod

Test Plan:
Click a link (in Safari and Chrome) and check if the app is opened. It works in debug mode - still have to check in release.

Update: tested it in release mode and it works.

Reviewers: kamil, inka

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8620