[native] Support the edit_role action on the keyserver
Summary:
This should be the final diff in the editing flow for custom roles, adding support on the keyserver to now support editing roles. It's important to note that since we pass the edited role name and
permissions, we needed something concrete to identify what role is being edited, so that's why there is a new parameter representing the existing role ID that should always exist if being edited (it's
optional since it's only defined in the editing flow, but not in the role creation flow).

[[ https://linear.app/comm/issue/ENG-4370/support-the-edit-role-action-on-the-keyserver | ENG-4370 ]]

Depends on D8524

Test Plan: Verified that editing roles actually successfully updates both the role names and the permissions

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8564