[web] Add an option to edit and delete select roles
Summary:
This diff introduces the dropdown select options to edit a role and delete a role. The implementation of editing roles will come in the next diff, and the implementation of deleting roles will come in the following one. Here, all roles besides `Admins` can be edited (we don't have a better way to check for Admins than by name for now, once [[
https://linear.app/comm/issue/ENG-4155/add-a-admin-role-column-for-threads | ENG-4155 ]] is taken on, that will be the better way). Similarly, all roles besides `Members` and `Admins` can be deleted.

Part of [[ https://linear.app/comm/issue/ENG-4429/add-an-option-to-edit-roles-for-all-roles-besides-admins | ENG-4429 ]] and [[
https://linear.app/comm/issue/ENG-4431/show-an-option-to-delete-roles-besides-admins-and-members | ENG-4431 ]].

Depends on D8597

Test Plan:
Confirmed that the options appear for each role as expected

{F653258}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ted, ashoat, tomek

Differential Revision: https://phab.comm.dev/D8623