[web] Push the create role modal with prepopulated fields when Edit Role is clicked
Summary:
This diff essentially handles the edit role workflow on web, where the permission enum options will be pre-selected if previously selected. The logic is similar to what was done in D8524 on `native`. The `keyserver` logic for handling editing roles is already done earlier in the stack

[[ https://linear.app/comm/issue/ENG-4430/push-the-create-role-modal-with-prepopulated-fields-when-edit-role-is | ENG-4430 ]]

Depends on D8623

Test Plan:
Verified that the modal is opened with the correct pre-populated fields after creating a new role then trying to edit it

{F653317}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8624