[web] Create the modals that will allow the user to confirm role deletion
Summary:
This should be the final stack in the entire roles project. This is the client-side logic for `web` to handle role deletion. Per the designs, the warning message is the same as on `native`, with
either a generic deletion message if nobody is assigned the role or a more specific one if members have this custom role.

[[ https://linear.app/comm/issue/ENG-4432/create-the-modals-that-will-allow-the-user-to-confirm-role-deletion | ENG-4432 ]]

Depends on D8624

Test Plan:
Verified that the modal appears as expected for roles with and without members, and pressing the delete button actually deletes

{F653374}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: tomek, ashoat

Differential Revision: https://phab.comm.dev/D8625