[keyserver] Make sure we don't call fetchServerThreadInfos with empty threadIDs
Summary:
This diff fixes a regression introduced in D8509. We should never specify an empty `threadIDs` parameter to the `FetchThreadInfosFilter` we pass to `fetchServerThreadInfos`.

This diff addresses an urgent regression – see [ENG-4529](https://linear.app/comm/issue/ENG-4529/sql-error-breaking-message-sends) for details.

Test Plan: I confirmed that created a message in a thread with an @-tag of somebody who is already in the thread does not cause the error described in the linked task

Reviewers: atul, tomek, michal, bartek, kamil, inka, ginsu, varun

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D8651