[native] Cache feature flags config
Summary:
Save fetched feature flags config in AsyncStorage. When the app starts, read the value from the storage and override it with a value from service when it is fetched.

This diff is an implementation of point 2 from https://linear.app/comm/issue/ENG-3252/improve-feature-flags-reliability

Test Plan: Ran the app and checked if after fresh start no value is read from the store and config is set after successful fetch. Then checked if restarting the app causes AsyncStorage read and saving to state which is then replaced by the config from service.

Reviewers: kamil, bartek