[lib] Rename `setUserAvatar` to `baseSetUserAvatar` in `*EditUserAvatarProvider`
Summary:
We want devs to to use the platform-specific `setUserAvatar` wrappers rather than interfacing with `setUserAvatar` directly... so renaming `setUserAvatar` to `baseSetUserAvatar` to make that more clear.

Also left a comment annotating the function, which at least in my editor shows a note on hover/when cursor is focused:

{F610023}

---

Depends on D8343

Test Plan: flow/eslint/searched codebase for "setUserAvatar".

Reviewers: ashoat, ginsu, rohan

Reviewed By: rohan

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8344