Move `native`-specific `uploadSelectedMedia` from `updateImageUserAvatar` to `nativeUpdateUserImageAvatar`
Summary:
We remove the `native`-specific `uploadSelectedMedia` call from `updateImageUserAvatar` and move it to `native`-specific `nativeUpdateUserImageAvatar`.

As of this diff, `updateImageUserAvatar` now takes `(imageAvatarUpdateRequest: ImageAvatarDBContent)` rather than the `native`-specific `(selection: NativeMediaSelection)`.

This diff also eliminates all usages of `useUploadSelectedMedia`, which means we can remove the prop in a subsequent diff at which point `BaseEditUserAvatarProvider` will be platform-agnostic! That'll let us consolidate `BaseEditUserAvatarProvider`, `WebEditUserAvatarProvider`, and `NativeEditUserAvatarProvider`.

---

Depends on D8345

Test Plan: Continue to be able to set image avatars on `native` as expected.

Reviewers: ashoat, ginsu, rohan

Reviewed By: ashoat, rohan

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8346