Consolidate `[Web/Native/Base]EditUserAvatarProvider` into `BaseEditUserAvatarProvider`
Summary:
As of this diff all of the platform-specific functionality has been moved out of `BaseEditUserAvatarProvider`, which allows us to consolidate the three previous providers into one.

This concludes most of the refactoring work aimed at decoupling `avatar-hooks` and `*AvatarProvider`s.

Next diff will involve renaming `BaseEditUserAvatarProvider` to `EditUserAvatarProvider` (was initially supposed to be part of this diff, but all the import changes and whatnot made it messy)

---

Depends on D8346

Test Plan: Still able to update avatars as expected on both `native` and `web`.

Reviewers: ashoat, ginsu, rohan

Reviewed By: rohan

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8347