[web] Move `exifRotate` from `input-state-container` to `media-utils`
Summary:
I needed to use `validateFile` outside of `input-state-container`, but saw that it took `exifRotate` as an argument.

I first considered moving `exifRotate` to `media-utils` and exporting it globally so I could access it from `input-state-container` and elsewhere. But upon reading the code further, it looks like we only ever use `exifRotate` in `processFile(file: File)`, so I just moved it right above without exporting.

---

Depends on D8363 (kind of)

Test Plan:
Logged value of `browser` and `exifRotate` in Safari/Chrome/FireFox:

{F611533}

{F611534}

{F611535}

Reviewers: ashoat, ginsu, rohan

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8364